Exported and renamed a few packages in the tellstick binding. See openhab/static-code-analysis#100
