+++ addons/binding/org.openhab.binding.tellstick/src/main/java/org/openhab/binding/tellstick/handler/TelldusDevicesHandler.java
@@ -29,8 +29,8 @@ import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.RefreshType;
 import org.eclipse.smarthome.core.types.State;
 import org.openhab.binding.tellstick.TellstickBindingConstants;
-import org.openhab.binding.tellstick.handler.live.xml.DataTypeValue;
-import org.openhab.binding.tellstick.handler.live.xml.TellstickNetSensor;
+import org.openhab.binding.tellstick.internal.live.xml.DataTypeValue;
+import org.openhab.binding.tellstick.internal.live.xml.TellstickNetSensor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.tellstick.device.TellstickException;
