+++ addons/binding/org.openhab.binding.tellstick/src/main/java/org/openhab/binding/tellstick/internal/live/TelldusLiveBridgeHandler.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.tellstick.handler.live;
+package org.openhab.binding.tellstick.internal.live;
 
 import java.util.List;
 import java.util.Vector;
@@ -22,16 +22,16 @@ import org.eclipse.smarthome.core.thing.binding.BaseBridgeHandler;
 import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.RefreshType;
 import org.eclipse.smarthome.core.types.State;
-import org.openhab.binding.tellstick.conf.TelldusLiveConfiguration;
 import org.openhab.binding.tellstick.handler.DeviceStatusListener;
 import org.openhab.binding.tellstick.handler.TelldusBridgeHandler;
 import org.openhab.binding.tellstick.handler.TelldusDeviceController;
 import org.openhab.binding.tellstick.handler.TelldusDevicesHandler;
-import org.openhab.binding.tellstick.handler.live.xml.DataTypeValue;
-import org.openhab.binding.tellstick.handler.live.xml.TellstickNetDevice;
-import org.openhab.binding.tellstick.handler.live.xml.TellstickNetDevices;
-import org.openhab.binding.tellstick.handler.live.xml.TellstickNetSensor;
-import org.openhab.binding.tellstick.handler.live.xml.TellstickNetSensors;
+import org.openhab.binding.tellstick.internal.conf.TelldusLiveConfiguration;
+import org.openhab.binding.tellstick.internal.live.xml.DataTypeValue;
+import org.openhab.binding.tellstick.internal.live.xml.TellstickNetDevice;
+import org.openhab.binding.tellstick.internal.live.xml.TellstickNetDevices;
+import org.openhab.binding.tellstick.internal.live.xml.TellstickNetSensor;
+import org.openhab.binding.tellstick.internal.live.xml.TellstickNetSensors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.tellstick.device.TellstickDeviceEvent;
